Add, update, delete players on the client side.

+++ client/js/presenters/playerstage.js
@@ -5,6 +5,7 @@ var $          = require('jquery')
   , Backbone   = require('backbone')
   , Marionette = require('backbone.marionette')
   , Stickit    = require('backbone.stickit')
+  , PlayerModel = require('../models/player')
   , Template   = require('../templates/playerstage.jade')
   , Mquc       = require('../mquc');
 
@@ -13,23 +14,23 @@ module.exports = Marionette.CompositeView.extend({
   template: Template,
 
   events: {
-    'click .btn-edit'   : 'toggleForm',
+    'click .btn-add'    : 'toggleAdd',
+    'click .btn-edit'   : 'toggleEdit',
     'click .btn-delete' : 'deletePlayer',
     'click .btn-save'   : 'sendForm'
   },
 
   ui: {
-    emptyStage    : '.stage-empty',
-    valueFixed    : '.value.fixed',
-    valueEditable : '.value.editable',
-    saveBtn       : '.btn-save'
+    fields  : 'input',
+    saveBtn : '.btn-save',
+    error   : '.form-error'
   },
 
   bindings: {
     '.input-name'     : 'name',
     '.input-fullname' : 'fullname',
     '.input-age'      : 'age',
-    '.input-jersey'   : 'jersey',
+    '.input-jersey'   : 'jersey_number',
     '.input-url'      : 'url'
   },
 
@@ -45,7 +46,7 @@ module.exports = Marionette.CompositeView.extend({
   onRender: function() {
     this.updateSelection();
     if (this.model) {
-      this.listenTo(this.model, 'change', this.enableSaveButton);
+      this.listenTo(this.model, 'change', this.toggleButton);
       this.stickit();
     }
   },
@@ -57,44 +58,85 @@ module.exports = Marionette.CompositeView.extend({
 
   updateSelection: function() {
     if (this.model) {
-      this.ui.emptyStage.addClass('hide');
       Mquc.vent.trigger('header:toggleSpinner', false);
     }
   },
 
-  toggleForm: function(e) {
+  toggleAdd: function(e) {
     e.preventDefault();
 
+    this.model = new PlayerModel();
+    this.render();
+    this.toggleEdit();
+  },
+
+  toggleEdit: function(e) {
+    if (e) e.preventDefault();
+
     if (!this.isEditMode()) {
-      this.ui.valueFixed.addClass('hide');
-      this.ui.valueEditable.removeClass('hide');
+      this.ui.fields.prop('readonly', false).removeClass('readonly');
+      this.ui.saveBtn.removeClass('hide');
     } else {
-      this.ui.valueEditable.addClass('hide');
-      this.ui.valueFixed.removeClass('hide');
+      this.ui.fields.prop('readonly', true).addClass('readonly');
+      this.ui.saveBtn.addClass('hide');
     }
   },
 
-  enableSaveButton: function() {
-    this.ui.saveBtn.prop('disabled', false);
+  toggleButton: function(enable) {
+    this.ui.saveBtn.prop('disabled', (enable ? false : true));
   },
 
   isEditMode: function() {
-    return !this.ui.valueEditable.first().hasClass('hide')
+    return !this.ui.fields.first().prop('readonly')
   },
 
   deletePlayer: function(e) {
     e.preventDefault();
-    Mquc.vent.trigger('header:toggleSpinner', true);
 
-    this.model.destroy({
-      success: function() {
-        Mquc.vent.trigger('header:toggleSpinner', false);
-      }
-    });
+    if (window.confirm('Do you want to delete this player?')) {
+      Mquc.vent.trigger('header:toggleSpinner', true);
+
+      this.model.destroy({
+        success: _.bind(function() {
+          Mquc.vent.trigger('header:toggleSpinner', false);
+          // TODO: Do not just clear the layout but close the view!
+          this.$el.html('');
+          Backbone.history.navigate('players');
+        }, this)
+      });
+    }
   },
 
   sendForm: function(e) {
-    e.preventDefault();
+    if (e) e.preventDefault();
+
+    Mquc.vent.trigger('header:toggleSpinner', true);
+    this.toggleButton(false);
+    this.ui.error.addClass('hide');
+
+    this.model.save({}, {
+      validate : true,
+      success  : _.bind(function() { this.afterSave(true) }, this),
+      error    : _.bind(function() { this.afterSave(false) }, this),
+    });
+
+    if (this.model.validationError) {
+      this.afterSave(false, this.model.validationError)
+    }
+  },
+
+  afterSave: function(isSuccessful, errorMsg) {
+    Mquc.vent.trigger('header:toggleSpinner', false);
+
+    if (isSuccessful) {
+      Mquc.vent.trigger('player:added', this.model);
+      Backbone.history.navigate('players/'+this.model.get('id'));
+      this.render();
+    } else {
+      errorMsg = errorMsg || 'Something went wrong.';
+      this.ui.error.html(errorMsg)
+                   .removeClass('hide');
+    }
   }
 
 });
